% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_publications.R
\name{get_closest}
\alias{get_closest}
\title{Match DiVA publications with Norwegian list on closest year}
\usage{
get_closest(pubs, norlist, matchpubs = "issn", matchlist = "issn")
}
\arguments{
\item{pubs}{publication list to match}

\item{norlist}{Norwegian list to match}

\item{matchpubs}{field to match on from pubs, default issn}

\item{matchlist}{field to match on from list, default issn}
}
\description{
Match DiVA publications with Norwegian list on closest year
}
