% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_publications.R
\name{match_diva_conference}
\alias{match_diva_conference}
\title{Match DiVA publications with Norwegian List journals by conference}
\usage{
match_diva_conference(journals, publications, closest_if_missing = FALSE)
}
\arguments{
\item{journals}{list of journal levels by year}

\item{publications}{publication list from DiVA}

\item{closest_if_missing}{set to TRUE to try closest year if level is missing}
}
\description{
Match DiVA publications with Norwegian List journals by conference
}
