% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_publications.R
\name{match_diva_publisher}
\alias{match_diva_publisher}
\title{Match DiVA publications with Norwegian List publisher}
\usage{
match_diva_publisher(publishers, publications, closest_if_missing = FALSE)
}
\arguments{
\item{publishers}{list of publisher levels}

\item{publications}{publication list from DiVA}

\item{closest_if_missing}{set to TRUE to try closest year if level is missing}
}
\description{
Match DiVA publications with Norwegian List publisher
}
