% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norwegian_list.R
\name{read_norwegian_list_parquet}
\alias{read_norwegian_list_parquet}
\title{Read parquet files norwegian_journals, norwegian_publishers}
\usage{
read_norwegian_list_parquet(location, from_s3 = FALSE)
}
\arguments{
\item{location}{a local path or an existing S3 bucket}

\item{from_s3}{set to TRUE to read from S3 storage (default FALSE)}
}
\description{
Read parquet files norwegian_journals, norwegian_publishers
}
