% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{separate_identificator}
\alias{separate_identificator}
\title{Separate delimited columns into rows according to separator}
\usage{
separate_identificator(pubs, identificator = "ISBN", keep = "PID", sep = ";")
}
\arguments{
\item{pubs}{publication list}

\item{identificator}{column to separate into rows (default "ISBN")}

\item{keep}{columns to keep in addition to identificator (default "PID")}

\item{sep}{the separator to use (default ";")}
}
\description{
Separate delimited columns into rows according to separator
}
