% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norwegian_list.R
\name{transpose_journals_list}
\alias{transpose_journals_list}
\title{Transpose Norwegian journals list}
\usage{
transpose_journals_list(journals_file)
}
\arguments{
\item{journals_file}{path to csv file with data on journals}
}
\value{
tibble
}
\description{
Transpose Norwegian journals list
}
