% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norwegian_list.R
\name{transpose_norwegian_list}
\alias{transpose_norwegian_list}
\title{Transpose Norwegian journals list}
\usage{
transpose_norwegian_list(journals_file, publishers_file)

transpose_norwegian_list(journals_file, publishers_file)

transpose_norwegian_list(journals_file, publishers_file)
}
\arguments{
\item{journals_file}{path to csv file with data on journals}

\item{publishers_file}{path to csv file with data on publishers}
}
\value{
tibble

tibble

list of 2 tibbles
}
\description{
Transpose Norwegian journals list

Transpose Norwegian publishers list

Transpose Norwegian lists of publication channels
}
