% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norwegian_list.R
\name{transpose_publishers_list}
\alias{transpose_publishers_list}
\title{Transpose Norwegian publishers list}
\usage{
transpose_publishers_list(publishers_file)
}
\arguments{
\item{publishers_file}{path to csv file with data on journals}
}
\value{
tibble
}
\description{
Transpose Norwegian publishers list
}
