% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norwegian_list.R
\name{write_norwegian_list_parquet}
\alias{write_norwegian_list_parquet}
\title{Write parquet files norwegian_journals, norwegian_publishers}
\usage{
write_norwegian_list_parquet(data, location, to_s3 = FALSE)
}
\arguments{
\item{data}{named list of 2 (journals, publishers)}

\item{location}{a local path or an existing S3 bucket}

\item{to_s3}{set to TRUE to write to S3 storage (default FALSE)}
}
\description{
Write parquet files norwegian_journals, norwegian_publishers
}
